Revert previous change.  After discussion with nathanw I understand that it is not necessary and was not the cause of the problems I was experiencing.

+++ lib/csu/mips/crt1.c
@@ -79,7 +79,7 @@ __start(char **ap,
 	char **argv;
 	char **env;
 
-	argc = * (int *) ap;
+	argc = * (long *) ap;
 	argv = ap + 1;
 	env  = ap + 2 + argc;
 	environ = env;
